Add UIAlertController extension
